add eslint; move tests to a separate dir
