+++ product-service/.eslintrc.js
@@ -1,0 +1,18 @@
+module.exports = {
+  env: {
+    browser: true,
+    es2021: true,
+    jest: true,
+  },
+  extends: [
+    'airbnb-base',
+  ],
+  parserOptions: {
+    ecmaVersion: 'latest',
+    sourceType: 'module',
+  },
+  ignorePatterns: ['**/*.config.js'],
+  rules: {
+    'import/prefer-default-export': 'off',
+  },
+};
