+++ product-service/constants/response.js
@@ -1,5 +1,5 @@
 export const RESPONSE_STATUSES = {
-    OK: 200,
-    NOT_FOUND: 404,
-    SERVER_ERROR: 500,
+  OK: 200,
+  NOT_FOUND: 404,
+  SERVER_ERROR: 500,
 };
