+++ product-service/handlers/getProductsById.js
@@ -1,41 +1,40 @@
 import products from '../mocks/products.json';
-import {ERRORS, ERROR_TYPES} from '../constants/error';
-import {RESPONSE_STATUSES} from "../constants/response";
-
+import { ERRORS, ERROR_TYPES } from '../constants/error';
+import { RESPONSE_STATUSES } from '../constants/response';
 
 const getProduct = (id = '') => new Promise((resolve, reject) => {
-    if (!id) {
-        reject(ERROR_TYPES.NO_ID);
-    }
+  if (!id) {
+    reject(ERROR_TYPES.NO_ID);
+  }
 
-    const product = products.find((item) => item.id === id);
+  const product = products.find((item) => item.id === id);
 
-    if (product) {
-        resolve(product);
-    } else {
-        reject(ERROR_TYPES.NOT_FOUND);
-    }
+  if (product) {
+    resolve(product);
+  } else {
+    reject(ERROR_TYPES.NOT_FOUND);
+  }
 });
 
 export const getProductsById = async (event = {}) => {
-    let response;
+  let response;
 
-    try {
-        const id = event.pathParameters && event.pathParameters.productId;
-        const product = await getProduct(id);
+  try {
+    const id = event.pathParameters && event.pathParameters.productId;
+    const product = await getProduct(id);
 
-        response = {
-            statusCode: RESPONSE_STATUSES.OK,
-            body: JSON.stringify(product),
-        };
-    } catch (err) {
-        const error = ERRORS[err] || ERRORS[ERROR_TYPES.DEFAULT];
+    response = {
+      statusCode: RESPONSE_STATUSES.OK,
+      body: JSON.stringify(product),
+    };
+  } catch (err) {
+    const error = ERRORS[err] || ERRORS[ERROR_TYPES.DEFAULT];
 
-        response = {
-            statusCode: error.code,
-            body: JSON.stringify(error),
-        };
-    }
+    response = {
+      statusCode: error.code,
+      body: JSON.stringify(error),
+    };
+  }
 
-    return response;
+  return response;
 };
