+++ product-service/package.json
@@ -6,13 +6,18 @@
     "@babel/core": "^7.18.6",
     "@babel/preset-env": "^7.18.6",
     "babel-loader": "^8.2.5",
+    "eslint": "^8.19.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.26.0",
     "jest": "^28.1.2",
     "serverless-webpack": "^5.7.1",
     "webpack": "^5.73.0"
   },
   "scripts": {
     "deploy": "sls deploy",
-    "test": "jest"
+    "test": "jest",
+    "lint": "eslint . --ext .js",
+    "lint:fix": "eslint . --ext .js --fix"
   },
   "author": "",
   "license": "ISC"
